requirements.txt
deepface==0.0.79
matplotlib==3.7.2
mediapipe==0.10.8
numpy==1.23.5
opencv_contrib_python==4.8.1.78
opencv_python==4.7.0.72
pandas==2.0.3
Pillow==9.5.0
Pillow==10.1.0
torch==2.1.0.dev20230622+cu118
torchvision==0.16.0.dev20230622+cu118
